+++ realm/realm-library/src/androidTest/java/io/realm/HandlerProxy.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.realm.proxy;
+package io.realm;
 
 import android.os.Handler;
 import android.os.Message;
@@ -24,7 +24,7 @@ import io.realm.HandlerController;
 /**
  * Handler decorator, to help intercept some messages before they are sent and received.
  */
-public abstract class HandlerProxy extends Handler {
+abstract class HandlerProxy extends Handler {
 
     private final HandlerController controller;